Release version 0.2.0: added support for creating and publishing dockerrun files version 2, added support for publishing multiple dockerrun files for the configured EC2 instance types

+++ src/main/scala/com/ovoenergy/sbt/ebs/ElasticBeanstalkPlugin.scala
@@ -1,5 +1,8 @@
 package com.ovoenergy.sbt.ebs
 
+import sbt._
+import Keys._
+
 import com.amazonaws.regions.{Regions, Region}
 import com.amazonaws.services.elasticbeanstalk.AWSElasticBeanstalkClient
 import com.amazonaws.services.elasticbeanstalk.model._
@@ -7,8 +10,6 @@ import com.amazonaws.services.s3.AmazonS3Client
 import com.amazonaws.services.s3.model._
 import com.typesafe.sbt.packager.NativePackagerKeys
 import com.typesafe.sbt.packager.docker.{DockerKeys, DockerPlugin}
-import sbt.Keys._
-import sbt._
 
 import scala.collection.JavaConversions._
 
@@ -17,32 +18,68 @@ object ElasticBeanstalkPlugin extends AutoPlugin with NativePackagerKeys with Do
   override def requires = DockerPlugin
 
   object autoImport {
+    lazy val ebsDockerrunVersion = settingKey[Int]("Version of the Dockerrun file (1 or 2)")
     lazy val ebsVersion = settingKey[String]("Version number of the application to create on Elastic Beanstalk")
     lazy val ebsS3Bucket = settingKey[String]("Name of the S3 bucket to publish docker configurations to")
     lazy val ebsS3AuthBucket = settingKey[String]("Name of the S3 bucket containing the docker auth config")
     lazy val ebsS3AuthKey = settingKey[String]("Name of the S3 file containing the docker auth config")
     lazy val ebsRegion = settingKey[Regions]("Region for the Elastic Beanstalk application and S3 bucket")
 
-    lazy val ebsStageDockerrunFile = taskKey[File]("Stages the Dockerrun.aws.json file")
-    lazy val ebsPublishDockerrunFile = taskKey[PutObjectResult]("Publishes the Dockerrun.aws.json file an S3 bucket")
-    lazy val ebsPublishAppVersion = taskKey[CreateApplicationVersionResult]("Publishes the new application version on Elastic Beanstalk")
+    lazy val ebsContainerPort = settingKey[Int]("Docker container port (Dockerrun version 1 only)")
+
+    lazy val ebsContainerMemory = settingKey[Int]("Memory required by the Docker container (Dockerrun version 2 only)")
+    lazy val ebsPortMappings = settingKey[Map[Int, Int]]("Port mappings for the Docker container (Dockerrun version 2 only)")
+    lazy val ebsEC2InstanceTypes = settingKey[Set[EC2InstanceType]]("EC2 instance types to generate Dockerrun files for (Dockerrun version 2 only)")
+
+    lazy val ebsStageDockerrunFiles = taskKey[List[File]]("Stages the Dockerrun.aws.json files")
+    lazy val ebsPublishDockerrunFiles = taskKey[List[PutObjectResult]]("Publishes the Dockerrun.aws.json files to an S3 bucket")
+    lazy val ebsPublishAppVersions = taskKey[List[CreateApplicationVersionResult]]("Publishes the application versions to Elastic Beanstalk")
+
+    val T2 = EC2InstanceTypes.T2
   }
   import autoImport._
 
   override lazy val projectSettings = Seq(
+    ebsDockerrunVersion := 1,
     ebsVersion := version.value,
     ebsS3AuthBucket := ebsS3Bucket.value,
     ebsS3AuthKey := ".dockercfg",
 
-    ebsStageDockerrunFile := {
-      val jsonFile = target.value / "aws" / "Dockerrun.aws.json"
-      jsonFile.delete()
-      IO.write(jsonFile, dockerrunFile.value)
-      jsonFile
+    ebsContainerPort := 8080,
+
+    ebsContainerMemory := 512,
+    ebsPortMappings := Map(
+      80 -> 8080,
+      8081 -> 8081
+    ),
+    ebsEC2InstanceTypes := Set.empty,
+
+    ebsStageDockerrunFiles := {
+
+      ebsDockerrunVersion.value match {
+        case 1 => List(DockerrunFileGenerator.generateDockerrunFileVersion1(
+          target.value, packageName.value, version.value,
+            dockerRepository.value, ebsS3AuthBucket.value, ebsS3AuthKey.value,
+            ebsContainerPort.value))
+        case 2 =>
+          if(ebsEC2InstanceTypes.value.isEmpty) {
+            List(DockerrunFileGenerator.generateDockerrunFileVersion2(
+              target.value, packageName.value, version.value,
+                dockerRepository.value, ebsS3AuthBucket.value, ebsS3AuthKey.value,
+                ebsPortMappings.value, Left(ebsContainerMemory.value)))
+          } else {
+            ebsEC2InstanceTypes.value.map { instanceType =>
+              DockerrunFileGenerator.generateDockerrunFileVersion2(
+                target.value, packageName.value, version.value,
+                  dockerRepository.value, ebsS3AuthBucket.value, ebsS3AuthKey.value,
+                  ebsPortMappings.value, Right(instanceType))
+            }.toList
+          }
+        case _ => throw new Exception("Invalid setting for 'ebsDockerrunVersion': must be 1 or 2")
+      }
     },
-    ebsPublishDockerrunFile := {
-      val key = s"${packageName.value}/${version.value}.json"
-      val jsonFile = ebsStageDockerrunFile.value
+    ebsPublishDockerrunFiles := {
+      val jsonFiles = ebsStageDockerrunFiles.value
 
       val bucket = ebsS3Bucket.value
 
@@ -54,51 +91,46 @@ object ElasticBeanstalkPlugin extends AutoPlugin with NativePackagerKeys with Do
         s3Client.createBucket(new CreateBucketRequest(bucket))
       }
 
-      s3Client.putObject(bucket, key, jsonFile)
+      jsonFiles.map { jsonFile =>
+        val key = s"${packageName.value}/${jsonFile.getName}"
+        if(s3Client.doesObjectExist(bucket, key) && !isSnapshot.value) {
+          throw new Exception(s"Unable to publish new dockerrun file to S3: file $key already exists")
+        }
+        s3Client.putObject(bucket, key, jsonFile)
+      }
     },
-    ebsPublishAppVersion := {
-      val key = s"${packageName.value}/${version.value}.json"
+    ebsPublishAppVersions := {
+
+      val versionsToPublish = ebsDockerrunVersion.value match {
+        case 1 => List(version.value)
+        case 2 =>
+          if(ebsEC2InstanceTypes.value.isEmpty) List(version.value)
+          else ebsEC2InstanceTypes.value.map { instanceType =>
+            s"${version.value}-${instanceType}"
+          }.toList
+        case _ => throw new Exception("Invalid setting for 'ebsDockerrunVersion': must be 1 or 2")
+      }
 
       val ebClient = new AWSElasticBeanstalkClient()
       ebClient.setRegion(Region.getRegion(ebsRegion.value))
 
       val applicationDescriptions = ebClient.describeApplications(new DescribeApplicationsRequest().withApplicationNames(packageName.value))
 
-      if (applicationDescriptions.getApplications.exists(_.getVersions contains version.value)) {
-        if(isSnapshot.value) ebClient.deleteApplicationVersion(new DeleteApplicationVersionRequest(packageName.value, version.value))
-        else throw new Exception(s"Unable to create new application version on Elastic Beanstalk: version ${version.value} already exists")
-      }
+      versionsToPublish.map { ebsVersion =>
+        if (applicationDescriptions.getApplications.exists(_.getVersions contains ebsVersion)) {
+          if(isSnapshot.value) ebClient.deleteApplicationVersion(new DeleteApplicationVersionRequest(packageName.value, ebsVersion))
+          else throw new Exception(s"Unable to create new application version on Elastic Beanstalk: version $ebsVersion already exists")
+        }
 
-      ebClient.createApplicationVersion(
-        new CreateApplicationVersionRequest()
-        .withApplicationName(packageName.value)
-        .withDescription(version.value)
-        .withVersionLabel(ebsVersion.value)
-        .withSourceBundle(new S3Location(ebsS3Bucket.value, key))
-      )
+        val key = s"${packageName.value}/${ebsVersion}.json"
+        ebClient.createApplicationVersion(
+          new CreateApplicationVersionRequest()
+          .withApplicationName(packageName.value)
+          .withDescription(version.value)
+          .withVersionLabel(ebsVersion)
+          .withSourceBundle(new S3Location(ebsS3Bucket.value, key))
+        )
+      }
     }
   )
-
-  lazy val dockerrunFile: Def.Initialize[String] = Def.setting {
-    val image = s"${packageName.value}:${version.value}"
-    val imageName = dockerRepository.value match {
-      case Some(repository) => s"$repository/$image"
-      case None => image
-    }
-    s"""|{
-        |  "AWSEBDockerrunVersion": "1",
-        |  "Image": {
-        |    "Name": "$imageName"
-        |  },
-        |  "Authentication": {
-        |    "Bucket": "${ebsS3AuthBucket.value}",
-        |    "Key": "${ebsS3AuthKey.value}"
-        |  },
-        |  "Ports": [
-        |    {
-        |      "ContainerPort": "8080"
-        |    }
-        |  ]
-        |}""".stripMargin
-  }
 }
